requirements.txt
Django==1.7.3
MySQL-python==1.2.5
django-social-auth==0.7.28
oauth2==1.5.211
oauthlib==0.7.2
python-openid==2.2.5
python-social-auth==0.2.6
requests==2.6.0
requests-oauthlib==0.4.2
django-render-json==0.2.6
shortuuid==0.4.2
pymemcache
